sage.misc.misc: Replace deprecated sorting 'uniq', move stable 'uniq' to sage.combinat.subset (with deprecation)
